misc/HtmlTestSuite/results/test-0926.xhtml

<?xml version="1.0" encoding="UTF-8"?><html xmlns="http://www.w3.org/1999/xhtml">
 <head>
  <title>Forms</title>
  <script src="/resources/testharness.js"/>
  <script src="/resources/testharnessreport.js"/>
 </head>
 <body>
  <p>
    </p><h3>button_validationMessage</h3>
  <p/>

  <hr/>

  <div id="log"/>

  <form action="" enctype="application/x-www-form-urlencoded" id="input_form" method="post">
    <p><button id="button_id">button</button></p>
  </form>
  <script>

    var button = document.getElementById("button_id");

    if (typeof(button.validationMessage) == "string") {
      test(function() {
        assert_equals(button.validationMessage, "", "validationMessage attribute is not correct.");
      });
    } else {
      test(function() {
        assert_unreached("validationMessage attribute is not exist.");
      });
    }

  </script>

 </body>
</html>
